Chess 23/05/2015
Got the game to a state were it reports "Check" when one player moves a
piece and causes the other player's king to be threatened. Once in
Check, the threatened player must move their King, and cannot move any
other pieces. Nearly done with this part, but at the moment the other
player can move pieces into a state of Check (if it is not already in
Check) - which is not quite right...